// services/proxy/src/models/Volo_Abp_Account_UpdateProfileDto.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type Volo_Abp_Account_UpdateProfileDto = {
  readonly extraProperties?: Record<string, any> | null;
  userName?: string | null;
  email?: string | null;
  name?: string | null;
  surname?: string | null;
  phoneNumber?: string | null;
  concurrencyStamp?: string | null;
};